Name insert done
Finished the name inseret part of the "game", tried to go with the "cursor" idea but realized 1hr+ into it that couldn't do it how i was doing it. So i did a "normal" and easy name insert (without the cursor). Maby if i hit my head enough times i could do it but i want to do it in a separate project.