fight thread
crated the thread for the fight phase. Now for the fight phase i just need to crate the acutal gameplay and moves. set all attributes private, not sure if useful but i read on stackoverflow that if i dont thay gon couse corruptions, vscode told me that there were some corruptions so i did it, it made things more comfortable for me too